Used javaexec to execute tests in forked jvm
When using Java libs there was an issue that the gradle daemon would
keep a JVM alive between test runs. This meant that additions to the
classloader from one run were being preserved for subsequent runs with
negative consequences.

Tests are now executed in a forked JVM using Gradle's javaexec task.
Longer term we could consider breaking the test runner into a separate
project, but inconsistencies in logging strategies might make this more
difficult than it first appears.
