+++ build.gradle
@@ -43,6 +43,8 @@ dependencies {
     implementation group: 'com.atlassian.commonmark', name: 'commonmark', version: '0.11.0'
     implementation group: 'com.atlassian.commonmark', name: 'commonmark-ext-gfm-tables', version: '0.9.0'
 
+    implementation group: 'com.xenomachina', name: 'kotlin-argparser', version: '2.0.7'
+
     testImplementation group: 'junit', name: 'junit', version: '4.12'
     testImplementation group: 'com.nhaarman', name: 'mockito-kotlin', version: '1.5.0'
